pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-slides-mvp"
version = "0.1.0"
description = "MVP pipeline scaffold for YouTube webinar slide extraction"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6.0,<3.0.0",
  "numpy>=1.26.0,<3.0.0",
  "Pillow>=10.0.0,<12.0.0",
  "PyMuPDF>=1.24.0,<2.0.0",
  "pytesseract>=0.3.10,<0.4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
